+++ src/hash_to_curve/map_g2.rs
@@ -521,7 +521,7 @@ fn test_osswu_semirandom() {
         0xe5,
     ]);
     for _ in 0..32 {
-        let input = Fp2::random(&mut rng);
+        let input = Fp2::try_from_rng(&mut rng).unwrap();
         let p = map_to_curve_simple_swu(&input);
         assert!(check_g2_prime(&p));