+++ src/fp6.rs
@@ -6,7 +6,7 @@ use core::ops::{Add, AddAssign, Mul, MulAssign, Neg, Sub, SubAssign};
 use subtle::{Choice, ConditionallySelectable, ConstantTimeEq, CtOption};
 
 #[cfg(feature = "pairings")]
-use rand_core::TryRngCore;
+use rand_core::TryRng;
 
 /// This represents an element $c_0 + c_1 v + c_2 v^2$ of $\mathbb{F}_{p^6} = \mathbb{F}_{p^2} / v^3 - u - 1$.
 pub struct Fp6 {
@@ -102,11 +102,11 @@ impl Fp6 {
     }
 
     #[cfg(feature = "pairings")]
-    pub(crate) fn try_from_rng<R: TryRngCore + ?Sized>(rng: &mut R) -> Result<Self, R::Error> {
+    pub(crate) fn try_random<R: TryRng + ?Sized>(rng: &mut R) -> Result<Self, R::Error> {
         Ok(Fp6 {
-            c0: Fp2::try_from_rng(rng)?,
-            c1: Fp2::try_from_rng(rng)?,
-            c2: Fp2::try_from_rng(rng)?,
+            c0: Fp2::try_random(rng)?,
+            c1: Fp2::try_random(rng)?,
+            c2: Fp2::try_random(rng)?,
         })
     }