+++ RELEASES.md
@@ -1,7 +1,8 @@
 # Unreleased
 ## Changed
 - MSRV is now 1.63.0.
-- Bumped dependencies to `digest 0.10`.
+- Bumped dependencies to `digest 0.10`, `ff 0.14`, `group 0.14`, `pairing 0.24`,
+  `rand_core 0.9`.
 
 # 0.8.0
 ## Changed